When cleaning up the reply, join lines with \n only
For now, this is a workaround for a problem where some plugins (LayerNames in
particular) do not handle `\r\n` as a separator well: it expects a single-char
separator.

With line endings canonicalized to `\r\n`, it was no longer possible to feed the
output of `focus-send` back to itself. So for the time being, until the firmware
is made smarter, canonicalized to `\n` instead.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>