Use a Trait to handle progress reporting
Rather than using function closures to handle progress reporting, use a Trait
instead. Include an implementation of the trait for `indicatif::ProgressBar` to
make the CLI tools even nicer.

This lets us get rid of the awkward `request_with_progress()` method, because
`request()` with an extra `None` argument is short enough to not need a separate
function.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>