bin: Slightly change the backup format
Add a `restore` array to the backed up data, which will be used during the
restore step. The idea is that `commands` may contain other data that we want to
save, but we don't want to restore.

While the `focus` tool will not make use of this, the plan is to share the
format between this tool, and Chrysalis, and Chrysalis will likely save a few
additional things, mostly for debugging purposes.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>