bin: Rearchitect how the binaries share code
Move all of the functionality of the binaries to a common struct, that
implements all of the commands. The binaries themselves are just a command line
parser and dispatcher layer on top of that now.

This makes the code a little cleaner, and easier to manage. It also removed a
whole lot of duplication.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>